changed most long descriptions to a MUCH cleaner & more readable syntax
Given the really messy nature of both the syntax and prose in general, I propose a new syntax convention:

If a block of writing is either long or contains many escape characters, the triple-quoted syntax should be used, and the quotes belong on their own line. The result is vastly more readable code, especially in cases of conversation texts, where there's quotes all over the place.